requirements.txt
fastapi
uvicorn
pywin32
requests
python-pptx
python-multipart
jinja2
